Add limits to the max num for the input

+++ sequencer/rpc_endpoint/src/rpc/types/codegen.rs
@@ -646,7 +646,7 @@ impl Transaction {
     ///
     /// A vector of bytes representing the transaction.
     /// if `send_fib` is true, then the execution represents fib(), otherwise it represents fact()
-    pub fn new_invoke_as_bytes(nonce: u64, arbitrary_calldata: u8, send_fib: bool) -> Vec<u8> {
+    pub fn new_invoke_as_bytes(nonce: u64, arbitrary_calldata: u16, send_fib: bool) -> Vec<u8> {
         let send_fib_felt = if send_fib {
             Felt252::new(0)
         } else {
